refactor: utill 파일명 단수로 변경 - #39
